Update jquery-ui to full 1.9.1, and include smoothness theme
This allows other apps to use a single version of jquery-ui that is
known to be compatible with Mezzanine's version of jquery.

This is jquery-ui-1.9.1.all, so all widgets are available and no
additional jquery code is needed. Also, the full smoothness theme
is included. Third-party apps may include other themes.