Fixed some encoding issues with set_cookie helper.
First, the %a and %b markers of strftime are locale sensitive and may
give non-English day / month abbreviations depending on the current
locale. Such localized abbreviations may contain non-ASCII characters
that trouble some web servers (e.g. gUnicorn throws a Unicode error).
It's fine to just let Django figure out proper values for expires and
max-age.

Secondly, Django silences CookieErrors raised when setting cookies
with keys with unsupported characters, which results in sending cookies
with None for their key and value.
It seems that the code that is responsible for this was introduced
with parsing of cookies received by server in mind (see ticket #13007),
so the legality of keys when setting new cookies may be enforced (to
help with forgetting to do some proper encoding of dynamic keys).

Thirdly, a few months ago Django introduced a workaround for cookie
values with Unicode characters in Python 2 (#19802). Included the
change for use with older Djangos.