Remove call to authenticate in ProfileForm.save()
I suspect the call to authenticate that I've taken out is be unnecessary - it's only called if it's a new signup, in which case ``password`` is non-blank and has just been passed to ``set_password``. This should always succeed and has no side-effects AFAICT - is there any need for it to be there?

I bring this up because I'm allowing password-less logons (i.e. via email only) by making a password optional at signup. Since the password isn't set, ``authenticate()`` returns None and the signup view breaks.

To make this case a bit more robust I've added a call to set_unusable_password() if the password is blank and it's a signup. This will only happen if the developer has modified the form to allow blank passwords in the first place.