Eval of lazy translations requires app registry
When Mezzanine patches project settings, some of the settings have
labels that are lazy translation objects. When register_setting()
checks to make sure that label exists and label is a lazy
translation object, it gets evaluated.

In Django 1.7, lazy translation objects cannot be evaluated until
the app registry is ready.