Updated to Chosen 1.1.0
I upgraded to Chosen 1.1.0 because the old version (0.9.5) was breaking
with newer versions of jQuery that don't include the browser property.
I chose to add a version suffix to the filenames in order to bust the
cache.  The new version of Chosen doesn't support the dark-select
styling that was previously applied in the header, so I removed those
stylings.