Updated to Chosen 0.9.12
I upgraded Chosen to 0.9.12 because the old version (0.9.5) was breaking
with newer versions of jQuery that don't include the browser property.
I chose to add a version suffix to the chosen JS filename in order to
bust the cache.  I didn't upgrade to the newest version of Chosen
(1.1.0 at the time of this commit) because it doesn't support the dark
styles that is used for the chosen select in the header.