Coded an override_locale context manager for use in tests.
The manager handles changing the locale in a thread-safe way and skips
the (whole) test in which it is used if the locale is not available.
With Python 3 may also be used as a decorator.

Note that for a locale to be available it must be installed for the
operating system. On *nixes you you may check what is available using
"locale -a" and for instance on Ubuntu add more locales using "apt-get
install language-pack-xx".