If keywords do not save allow admin to force save
If keywords fail to submit for some reason clicking any of the save buttons on an admin page will appear to do nothing.  This commit allows the admin user to choose whether or not to continue saving.  It also makes debugging easier as it makes apparent that there is a problem and what the problem is with.