Correctly handle custom email fields in mezzanine.forms
The `is_a` method checks whether a field is one of Mezzanine's built-in
form fields. As far as I can tell, every other usage is for the purpose
of initializing the built-in fields and widgets, so it makes sense to
exclude user-defined fields.

However in this instance, we only want to know whether the field is an
EmailField, not that it is *the* built-in Mezzanine email field.
Therefore, check the class rather than identity.

The effective change here is that user-defined fields (in
FORMS_EXTRA_FIELDS) which subclass django's EmailField will be able to
receive confirmation emails.