Add check on default_storage for thumbnail
This improve load times on the media library and on pages that use thumbnail often by up to 20x on storages like Amazon S3. The issue is that often the thumbnail has already been made but isn't on the local system because we are using ephemeral storage locally. Without this check we often end up recreating the thumbnail that already exists of the ephemeral storage has been wiped recently like on Heroku for when you push a new update, adding the check saves a ton of processing time.