utils: get all models for a given app
Add a lazy function that retrieves all models for a given app,
identified by it's `app_label`.
This makes it possible to include all models for an app in
ADMIN_MENU_ORDER setting, so that it's at the right spot and no models
are "missing" (moved down) when upgrading the dependency.

Usage:

    ADMIN_MENU_ORDER = (
    (
        "Content", (
            "pages.Page",
            "blog.BlogPost",
            "generic.ThreadedComment",
            (_("Media Library"), "media-library"),)
    ),
    # ... more here ...
    (
        _("Social Accounts"), get_models_for_app('socialaccount')
    ),
    # ... more here ...

A tuple of classes can also be returned, which may or may not prove to
be useful in the future.